extract graph sampler from trajectory balance module
